+++ profile/libs/general/activated.php
@@ -20,11 +20,11 @@
               <tr>
                 <td><?php echo esc_html_x("Profile Logo Image","profile-section", "peprodev-ups");?></td>
                 <td>
-                  <input type="hidden" data-id="<?php echo esc_attr( $this->read("logo-id","") );?>" id="profile-section-logo"
-                  value="<?php echo esc_attr( $this->read("logo", $this->icon) );?>"
+                  <input type="hidden" data-id="<?php echo esc_attr( $this->read("custom_logo_id","") );?>" id="profile-section-logo"
+                  value="<?php echo esc_attr( $this->read("custom_logo", $this->icon) );?>"
                   class="form-control primary" placeholder="<?php echo esc_html_x("Logo URL","profile-section","peprodev-ups");?>" />
                   <div class="flex flex-sb">
-                    <img style="border-radius: 5px;" src="<?php echo esc_attr( $this->read("logo","") );?>" id="profile-img" width="86px"/>
+                    <img style="border-radius: 5px;" src="<?php echo esc_attr( $this->read("custom_logo","") );?>" id="profile-img" width="86px"/>
                     <button type="button" id="selectlogoimg" style="padding: 12px; display: block;" class="btn btn-primary mediapicker icn-btn"
                     data-ref="#profile-section-logo"
                     data-ref4="#profile-img"
@@ -90,7 +90,7 @@
                   data-text-off='<?php echo esc_html_x("No, Hide it","profile-section", "peprodev-ups");?>'
                   data-on='check_box'
                   data-off='check_box_outline_blank'
-                  data-checked='<?php echo esc_attr( $this->read("showwelcome","true") === "true" ? "true" : "false" );?>' id="showwelcome"></a>
+                  data-checked='<?php echo esc_attr( $this->read("show_welcome","true") === "true" ? "true" : "false" );?>' id="show_welcome"></a>
                 </td>
               </tr>
               <?php if ($this->_wc_activated()){ ?>
@@ -103,7 +103,7 @@
                     data-text-off='<?php echo esc_html_x("No, Hide it","profile-section", "peprodev-ups");?>'
                     data-on='check_box'
                     data-off='check_box_outline_blank'
-                    data-checked='<?php echo esc_attr( $this->read("woocommercestats","true") === "true" ? "true" : "false" );?>' id="woocommercestats"></a>
+                    data-checked='<?php echo esc_attr( $this->read("woocommerce_stats","true") === "true" ? "true" : "false" );?>' id="woocommercestats"></a>
                   </td>
                 </tr>
                 <!-- Show WooCommerce Orders -->
@@ -115,7 +115,7 @@
                     data-text-off='<?php echo esc_html_x("No, Hide it","profile-section", "peprodev-ups");?>'
                     data-on='check_box'
                     data-off='check_box_outline_blank'
-                    data-checked='<?php echo esc_attr( $this->read("woocommerceorders","true") === "true" ? "true" : "false" );?>' id="woocommerceorders"></a>
+                    data-checked='<?php echo esc_attr( $this->read("woocommerce_orders","true") === "true" ? "true" : "false" );?>' id="woocommerceorders"></a>
                   </td>
                 </tr>
               <?php } ?>
@@ -129,17 +129,17 @@
                   data-on='check_box'
                   data-off='check_box_outline_blank'
                   data-togglel='[showoncustomcontent]'
-                  data-checked='<?php echo esc_attr( $this->read("showcustomtext","true") === "true" ? "true" : "false" );?>' id="showcustomtext"></a>
+                  data-checked='<?php echo esc_attr( $this->read("show_custom_text","true") === "true" ? "true" : "false" );?>' id="showcustomtext"></a>
                 </td>
               </tr>
               <tr showoncustomcontent>
                 <td><?php echo esc_html_x("Content Position","profile-section", "peprodev-ups");?></td>
                 <td>
                   <select id="customposition">
-                    <?php echo "<option " . selected($this->read("customposition","p2"),"p1",false)." value='p1'>" . esc_html_x("Before Welcome message","profile-section","peprodev-ups") . "</option>";?>
-                    <?php echo "<option " . selected($this->read("customposition","p2"),"p2",false)." value='p2'>" . esc_html_x("After Welcome message","profile-section","peprodev-ups") . "</option>";?>
-                    <?php echo "<option " . selected($this->read("customposition","p2"),"p3",false)." value='p3'>" . esc_html_x("After WooCommerce Stats","profile-section","peprodev-ups") . "</option>";?>
-                    <?php echo "<option " . selected($this->read("customposition","p2"),"p4",false)." value='p4'>" . esc_html_x("After WooCommerce Orders","profile-section","peprodev-ups") . "</option>";?>
+                    <?php echo "<option " . selected($this->read("custom_position","p2"),"p1",false)." value='p1'>" . esc_html_x("Before Welcome message","profile-section","peprodev-ups") . "</option>";?>
+                    <?php echo "<option " . selected($this->read("custom_position","p2"),"p2",false)." value='p2'>" . esc_html_x("After Welcome message","profile-section","peprodev-ups") . "</option>";?>
+                    <?php echo "<option " . selected($this->read("custom_position","p2"),"p3",false)." value='p3'>" . esc_html_x("After WooCommerce Stats","profile-section","peprodev-ups") . "</option>";?>
+                    <?php echo "<option " . selected($this->read("custom_position","p2"),"p4",false)." value='p4'>" . esc_html_x("After WooCommerce Orders","profile-section","peprodev-ups") . "</option>";?>
                   </select>
                 </td>
               </tr>
@@ -147,7 +147,7 @@
                   <td colspan="2">
                     <?php
                       wp_editor(
-                          stripslashes($this->read("customhtml")),
+                          stripslashes($this->read("custom_html")),
                           "peprodev-ups-customhtml",
                           array(
                           'textarea_name'    => "customhtml",
@@ -192,15 +192,15 @@
               <tr>
                 <td colspan="2">
                   <?php echo esc_html_x("Global CSS","profile-section", "peprodev-ups");?>
-                  <textarea name="css" class="codeditor" id="csseditor" spellcheck="false" dir="ltr" rows="8" cols="80"><?php echo html_entity_decode(stripslashes($this->read("css")));?></textarea>
-                  <textarea name="css" style="display:none !important;" class="codeditor" id="css" spellcheck="false" dir="ltr" rows="8" cols="80"><?php echo html_entity_decode(stripslashes($this->read("css")));?></textarea>
+                  <textarea name="css" class="codeditor" id="csseditor" spellcheck="false" dir="ltr" rows="8" cols="80"><?php echo html_entity_decode(stripslashes($this->read("custom_css")));?></textarea>
+                  <textarea name="css" style="display:none !important;" class="codeditor" id="css" spellcheck="false" dir="ltr" rows="8" cols="80"><?php echo html_entity_decode(stripslashes($this->read("custom_css")));?></textarea>
                 </td>
               </tr>
               <tr>
                 <td colspan="2">
                   <?php echo esc_html_x("Global JS","profile-section", "peprodev-ups");?>
-                  <textarea name="js" class="codeditor" id="jseditor" spellcheck="false" dir="ltr" rows="8" cols="80"><?php echo html_entity_decode(stripslashes($this->read("js")));?></textarea>
-                  <textarea name="js" style="display:none !important;" class="codeditor" id="js" spellcheck="false" dir="ltr" rows="8" cols="80"><?php echo html_entity_decode(stripslashes($this->read("js")));?></textarea>
+                  <textarea name="js" class="codeditor" id="jseditor" spellcheck="false" dir="ltr" rows="8" cols="80"><?php echo html_entity_decode(stripslashes($this->read("custom_js")));?></textarea>
+                  <textarea name="js" style="display:none !important;" class="codeditor" id="js" spellcheck="false" dir="ltr" rows="8" cols="80"><?php echo html_entity_decode(stripslashes($this->read("custom_js")));?></textarea>
                 </td>
               </tr>
           </tbody>