+++ changelog.md
@@ -8,6 +8,7 @@ Developed by [Pepro Development Group](https://pepro.dev/), Lead Developer: [Ami
 <summary><strong>Version 8.x.x</strong></summary>
 &nbsp;
 
+- [Version 8.0.3](#version-803)
 - [Version 8.0.2](#version-802)
 - [Version 8.0.1](#version-801)
 - [Version 8.0.0](#version-800)
@@ -68,6 +69,11 @@ Developed by [Pepro Development Group](https://pepro.dev/), Lead Developer: [Ami
 </details>
 
 ---
+
+## Version 8.0.3
+- Release date: 2025-05-31 | 1404-03-10  [&uarr;](#table-of-contents)
+- **Fixed** some Backend Setting UI issues
+- **Fixed** Custom text on dashboard not rendered properly
 
 ## Version 8.0.2
 - Release date: 2025-05-28 | 1404-03-07  [&uarr;](#table-of-contents)